Transfer the domain to DNS
Gandi has been brought by a PE firm who look to be destroying the
value of it. I've tried OVH, but they don't support root ALIAS, ANAME,
or CNAME and hence can't be used with this infrastructure. Instead
Cloudflare can so I'll switch to them.

Note that the .dev TLD is not yet supported by cloudflare, when it is
this will be a full transfer.